Clean up CTAS benchmark result handling
