+++ presto/docker/config/template/etc_common/catalog/hive_output.properties
@@ -2,7 +2,7 @@
 # mounted separately from PRESTO_DATA_DIR so source data can remain read-only.
 connector.name=hive-hadoop2
 hive.metastore=file
-hive.metastore.catalog.dir=file:/var/lib/presto/data/hive/benchmark_output
+hive.metastore.catalog.dir=file:/var/lib/presto/data/hive/ctas_scratch_output
 hive.allow-drop-table=true
 # Native workers otherwise stage writes under their container-private /tmp,
 # which prevents the coordinator from committing the files. Write directly to
